Remodeled file hierarchy, code cleanup
Also started work on a GUI for administrator controls, but I need to understand asynchronous programming better before I can properly implement any of the ideas I have :(

Using a command in the Discord client to properly close the bot or abruptly ending the process is unpractical, so implementing a kill switch in the GUI will probably be the first move.

I'm also thinking about having this bot be able to run from a single executable, since starting it from a script nested within a folder is kind of awkward. Not sure how that'll work exactly yet, but I've already started orienting some parts of the code around the possibility.