Scrapped graphic interface
After some researching, I realize that implementing a GUI that runs parallel to the bot would be very difficult to implement and it's something the Discord library was not designed to be compatible with. Simply closing the bot console is enough for killing the bot, and any administrator commands can simply be locked behind a check.